# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: CIDB Tender API
  version: 1.0.0
  description: "Standalone normalized API over publicly available CIDB tender
    information. Authenticate with the `X-API-Key` header. Source: Construction
    Industry Development Board (CIDB)."
components:
  securitySchemes:
    apiKey:
      type: apiKey
      name: X-API-Key
      in: header
      description: API key issued for this service
...
paths:
...
  /tenders:
    get:
      tags:
        - tenders
      description: List tenders with search, filters, sorting and pagination.
      parameters:
        - schema:
            type: integer
            minimum: 1
            default: 1
          in: query
          name: page
          required: false
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
          in: query
          name: limit
          required: false
        - schema:
            type: string
            maxLength: 200
          in: query
          name: search
          required: false
        - schema:
            type: string
            enum:
              - OPEN
              - CLOSING_SOON
              - CLOSED
              - CANCELLED
              - AWARDED
              - ARCHIVED
              - UNKNOWN
          in: query
          name: status
          required: false
        - schema:
            type: string
            maxLength: 100
          in: query
          name: province
          required: false
        - schema:
            type: string
            maxLength: 20
          in: query
          name: cidbGrade
          required: false
        - schema:
            type: string
            maxLength: 20
          in: query
          name: cidbClass
          required: false
        - schema:
            type: string
            maxLength: 200
          in: query
          name: organisation
          required: false
        - schema:
            type: string
          in: query
          name: publishedFrom
          required: false
        - schema:
            type: string
          in: query
          name: publishedTo
          required: false
        - schema:
            type: string
          in: query
          name: closingFrom
          required: false
        - schema:
            type: string
          in: query
          name: closingTo
          required: false
        - schema:
            type: string
            enum:
              - publishedDate
              - closingDate
              - createdAt
              - updatedAt
              - bidNumber
              - title
            default: publishedDate
          in: query
          name: sort
          required: false
        - schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
          in: query
          name: order
          required: false
      security:
        - apiKey: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        source:
                          type: string
                        externalId:
                          type: string
                        bidNumber:
                          type: string
                          nullable: true
                        title:
                          type: string
                        description:
                          type: string
                          nullable: true
                        organisation:
                          type: string
                          nullable: true
                        province:
                          type: string
                          nullable: true
                        location:
                          type: string
                          nullable: true
                        municipality:
                          type: string
                          nullable: true
                        tenderType:
                          type: string
                          nullable: true
                        status:
                          type: string
                          enum:
                            - OPEN
                            - CLOSING_SOON
                            - CLOSED
                            - CANCELLED
                            - AWARDED
                            - ARCHIVED
                            - UNKNOWN
                        publishedDate:
                          type: string
                          nullable: true
                        closingDate:
                          type: string
                          nullable: true
                        briefingDate:
                          type: string
                          nullable: true
                        briefingRequired:
                          type: boolean
                        briefingLocation:
                          type: string
                          nullable: true
                        cidbGrade:
                          type: string
                          nullable: true
                        cidbClass:
                          type: array
                          items:
                            type: string
                        estimatedValue:
                          type: string
                          nullable: true
                        contact:
                          type: object
                          properties:
                            name:
                              type: string
                              nullable: true
                            email:
                              type: string
                              nullable: true
                            phone:
                              type: string
                              nullable: true
                          required:
                            - name
                            - email
                            - phone
                          additionalProperties: false
                        documents:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                              name:
                                type: string
                              documentType:
                                type: string
                              url:
                                type: string
                              fileName:
                                type: string
                                nullable: true
                              mimeType:
                                type: string
                                nullable: true
                            required:
                              - id
                              - name
                              - documentType
                              - url
                              - fileName
                              - mimeType
                            additionalProperties: false
                        sourceUrl:
                          type: string
                        firstSeenAt:
                          type: string
                        lastSeenAt:
                          type: string
                        createdAt:
                          type: string
                        updatedAt:
                          type: string
                      required:
                        - id
                        - source
                        - externalId
                        - bidNumber
                        - title
                        - description
                        - organisation
                        - province
                        - location
                        - municipality
                        - tenderType
                        - status
                        - publishedDate
                        - closingDate
                        - briefingDate
                        - briefingRequired
                        - briefingLocation
                        - cidbGrade
                        - cidbClass
                        - estimatedValue
                        - contact
                        - documents
                        - sourceUrl
                        - firstSeenAt
                        - lastSeenAt
                        - createdAt
                        - updatedAt
                      additionalProperties: false
                  pagination:
                    type: object
                    properties:
                      page:
                        type: number
                      limit:
                        type: number
                      total:
                        type: number
                      totalPages:
                        type: number
                    required:
                      - page
                      - limit
                      - total
                      - totalPages
                    additionalProperties: false
                required:
                  - data
                  - pagination
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
  /tenders/search:
    get:
      tags:
        - tenders
      description: Full-text style search across bid number, title, description,
        organisation and location.
      parameters:
        - schema:
            type: string
            minLength: 1
            maxLength: 200
          in: query
          name: q
          required: true
        - schema:
            type: integer
            minimum: 1
            default: 1
          in: query
          name: page
          required: false
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
          in: query
          name: limit
          required: false
      security:
        - apiKey: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        source:
                          type: string
                        externalId:
                          type: string
                        bidNumber:
                          type: string
                          nullable: true
                        title:
                          type: string
                        description:
                          type: string
                          nullable: true
                        organisation:
                          type: string
                          nullable: true
                        province:
                          type: string
                          nullable: true
                        location:
                          type: string
                          nullable: true
                        municipality:
                          type: string
                          nullable: true
                        tenderType:
                          type: string
                          nullable: true
                        status:
                          type: string
                          enum:
                            - OPEN
                            - CLOSING_SOON
                            - CLOSED
                            - CANCELLED
                            - AWARDED
                            - ARCHIVED
                            - UNKNOWN
                        publishedDate:
                          type: string
                          nullable: true
                        closingDate:
                          type: string
                          nullable: true
                        briefingDate:
                          type: string
                          nullable: true
                        briefingRequired:
                          type: boolean
                        briefingLocation:
                          type: string
                          nullable: true
                        cidbGrade:
                          type: string
                          nullable: true
                        cidbClass:
                          type: array
                          items:
                            type: string
                        estimatedValue:
                          type: string
                          nullable: true
                        contact:
                          type: object
                          properties:
                            name:
                              type: string
                              nullable: true
                            email:
                              type: string
                              nullable: true
                            phone:
                              type: string
                              nullable: true
                          required:
                            - name
                            - email
                            - phone
                          additionalProperties: false
                        documents:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                              name:
                                type: string
                              documentType:
                                type: string
                              url:
                                type: string
                              fileName:
                                type: string
                                nullable: true
                              mimeType:
                                type: string
                                nullable: true
                            required:
                              - id
                              - name
                              - documentType
                              - url
                              - fileName
                              - mimeType
                            additionalProperties: false
                        sourceUrl:
                          type: string
                        firstSeenAt:
                          type: string
                        lastSeenAt:
                          type: string
                        createdAt:
                          type: string
                        updatedAt:
                          type: string
                      required:
                        - id
                        - source
                        - externalId
                        - bidNumber
                        - title
                        - description
                        - organisation
                        - province
                        - location
                        - municipality
                        - tenderType
                        - status
                        - publishedDate
                        - closingDate
                        - briefingDate
                        - briefingRequired
                        - briefingLocation
                        - cidbGrade
                        - cidbClass
                        - estimatedValue
                        - contact
                        - documents
                        - sourceUrl
                        - firstSeenAt
                        - lastSeenAt
                        - createdAt
                        - updatedAt
                      additionalProperties: false
                  pagination:
                    type: object
                    properties:
                      page:
                        type: number
                      limit:
                        type: number
                      total:
                        type: number
                      totalPages:
                        type: number
                    required:
                      - page
                      - limit
                      - total
                      - totalPages
                    additionalProperties: false
                required:
                  - data
                  - pagination
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
...
  /tenders/{id}:
    get:
      tags:
        - tenders
      description: Full tender detail including documents. Accepts the internal id or
        the stable externalId.
      parameters:
        - schema:
            type: string
            minLength: 1
            maxLength: 100
          in: path
          name: id
          required: true
      security:
        - apiKey: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  source:
                    type: string
                  externalId:
                    type: string
                  bidNumber:
                    type: string
                    nullable: true
                  title:
                    type: string
                  description:
                    type: string
                    nullable: true
                  organisation:
                    type: string
                    nullable: true
                  province:
                    type: string
                    nullable: true
                  location:
                    type: string
                    nullable: true
                  municipality:
                    type: string
                    nullable: true
                  tenderType:
                    type: string
                    nullable: true
                  status:
                    type: string
                    enum:
                      - OPEN
                      - CLOSING_SOON
                      - CLOSED
                      - CANCELLED
                      - AWARDED
                      - ARCHIVED
                      - UNKNOWN
                  publishedDate:
                    type: string
                    nullable: true
                  closingDate:
                    type: string
                    nullable: true
                  briefingDate:
                    type: string
                    nullable: true
                  briefingRequired:
                    type: boolean
                  briefingLocation:
                    type: string
                    nullable: true
                  cidbGrade:
                    type: string
                    nullable: true
                  cidbClass:
                    type: array
                    items:
                      type: string
                  estimatedValue:
                    type: string
                    nullable: true
                  contact:
                    type: object
                    properties:
                      name:
                        type: string
                        nullable: true
                      email:
                        type: string
                        nullable: true
                      phone:
                        type: string
                        nullable: true
                    required:
                      - name
                      - email
                      - phone
                    additionalProperties: false
                  documents:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        documentType:
                          type: string
                        url:
                          type: string
                        fileName:
                          type: string
                          nullable: true
                        mimeType:
                          type: string
                          nullable: true
                      required:
                        - id
                        - name
                        - documentType
                        - url
                        - fileName
                        - mimeType
                      additionalProperties: false
                  sourceUrl:
                    type: string
                  firstSeenAt:
                    type: string
                  lastSeenAt:
                    type: string
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
                required:
                  - id
                  - source
                  - externalId
                  - bidNumber
                  - title
                  - description
                  - organisation
                  - province
                  - location
                  - municipality
                  - tenderType
                  - status
                  - publishedDate
                  - closingDate
                  - briefingDate
                  - briefingRequired
                  - briefingLocation
                  - cidbGrade
                  - cidbClass
                  - estimatedValue
                  - contact
                  - documents
                  - sourceUrl
                  - firstSeenAt
                  - lastSeenAt
                  - createdAt
                  - updatedAt
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
...
servers:
  - url: /api/v1
    description: Current server (v1)
...